Support systems with 128-bit pointers.
In order to allow meaningful comparisons between 32- and 64-bit systems,
core_list_init hardcodes the size of pointers to be 8 bytes.  This
controls the number list elements allocated for a buffer of a given
size.  On a system with 128-bit pointers this results in data corruption
as the list data is written over list_head objects.

Address this by adding a POINTER_SPACE define which can be changed to
16.  Update the checksums in the known_crc arrays for the run1.log and
run2.log cases.